manage wrapper and AXIResponder, step 4: disable embedded spm partial get, add tags for SPM, use LRU replacement. Always replacing spm.begin() will lead to wrong answer in mod_googlenet_conv2_3x3_int16, but using LRU gives correct answer.

+++ ext/rtl/model_nvdla/axiResponder.hh
@@ -100,7 +100,7 @@ private:
     // dma & spm
     // function together with inflight_req & inflight_req_order
     std::map<uint64_t, uint32_t> inflight_dma_addr_size;    // record the inflight dma request sizes
-    std::queue<uint32_t> inflight_dma_addr_queue;           // keep dma request order
+    std::queue<uint64_t> inflight_dma_addr_queue;           // keep dma request order
 
     // prefetch
     uint32_t pft_threshold;
